requirements.txt
lib_ocr
google-generativeai
python-dotenv
pytesseract
Pillow
opencv-python
pyautogui
